[Fix] 대시보드 카드아이템 무한스크롤 실제 데이터에 적용
