+++ src/app/api/dashboard/columnList/route.ts
@@ -10,17 +10,18 @@ export const GET = async (req: Request) => {
   const { searchParams } = new URL(req.url);
   const cursorId = searchParams.get("cursorId");
   const columnId = searchParams.get("columnId");
+  const size = searchParams.get("size");
   
   if (!token) {
     return new NextResponse("사용자 정보를 찾을 수 없습니다.", { status: 401 });
   }
 
-  if (!cursorId || !columnId) {
+  if (!cursorId || !columnId || !size) {
     return new NextResponse("컬럼의 카드 정보를 가져오는 데이터를 확인해주세요.", { status: 400 });
   }
 
   try {
-    const response = await apiClient.get(`/cards?size=10&cursorId=${cursorId}&columnId=${columnId}`, {
+    const response = await apiClient.get(`/cards?size=${size}&cursorId=${cursorId}&columnId=${columnId}`, {
       headers: {
         Authorization: `Bearer ${token}`
       }
