+++ src/app/dashboard/components/ColumnList.tsx
@@ -16,6 +16,7 @@ const ColumnList = ({ columnTitle, columnId }: IProps) => {
   const [cardList, setCardList] = useState<ICard[]>([]);
   const [cursorId, setCursorId] = useState<number>(1);
   const [hasMore, setHasMore] = useState(true);
+  const [size, setSize] = useState(3);
   const observeRef = useRef<IntersectionObserver | null>(null);
   const loadingRef = useRef<HTMLDivElement | null>(null);
 
@@ -23,20 +24,21 @@ const ColumnList = ({ columnTitle, columnId }: IProps) => {
     if (!hasMore) return;
 
     try {
-      const response = await axios.get(`/api/dashboard/columnList?cursorId=${cursorId}&columnId=${columnId}`);
+      const response = await axios.get(`/api/dashboard/columnList?cursorId=${cursorId}&columnId=${columnId}&size=${size}`);
 
       if (response.status === 200) {
         setCardList((prev) => [...prev, ...response.data.cards]);
         setCursorId(response.data.cursorId);
       }
 
-      if (response.data.cards.length < 10) {
+      if (response.data.cards.length < size) {
         setHasMore(false);
+        toast.success("더 가져올 카드가 없습니다.");
       }
     } catch (error) {
       if (axios.isAxiosError(error)) {
         console.error("ColumnList getCardList에서 api 오류 발생", error);
-        toast.error(error.response?.data.message);
+        toast.error(error.response?.data);
       }
     }
   };
@@ -62,15 +64,14 @@ const ColumnList = ({ columnTitle, columnId }: IProps) => {
         observeRef.current?.unobserve(loadingRef.current);
       }
     };
-  }, [hasMore]);
+  }, [hasMore, cursorId, size]);
 
   const handleAddTodo = () => {
     // 모달 만들어지면 모달 연결
-    console.log("모달 오픈");
   };
 
   const handleEditModal = () => {
-
+    // 모달 만들어지면 모달 연결
   }
 
   return (
